requirements.txt
Flask
python-dotenv
streamlit
Pillow 
google-generativeai